Fixed links in README.
As the repository has been forked to it's new location, the links had to be
adjusted. Most of the links will point to the new fork now, but the links of the
screenshots can't be adjusted at the moment, as GitHub cloud has been replaced
by GitHub releases, which don't support individual assets at the moment. The
coresponding issue has been updated to figure out a better solution.